Reflection
For my code I first looked at the book and it gave me a code on how to move the the mouse but it was a circle so I used some of that code but changed it in to square. I used the book and p5.org to look up. I struggled with changing the circle to square. I really struggled so I asked a tutor and Jesse helped me out. I didn't have any plan.